pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cancersd"
version = "0.1.0"
description = "CancerSD: Cancer subtype diagnosis under incomplete multi-omics data"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Yongqi Bu" }
]
dependencies = [
  "numpy>=1.23",
  "pandas>=1.5",
  "scikit-learn>=1.2",
  "pyyaml>=6.0",
  "torch>=2.0",
  "tqdm>=4.64",
  "matplotlib>=3.6",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
cancersd-train = ""
cancersd-eval = ""
cancersd-preprocess-tcga = ""
cancersd-preprocess-geo = ""